Slightly less evil global
