+++ spex_simulator/include/Gui.h
@@ -1,7 +1,7 @@
 #ifndef GUI_H
 #define GUI_H
 
-#include "Arduino.h"
+#include "PinState.h"
 #include "imgui.h"
 #include "imgui_impl_glfw.h"
 #include "imgui_impl_opengl3.h"
